Philips-backend-main/requirements.txt
Flask
flask-cors
flask-sqlalchemy
python-dotenv
psycopg2-binary
transformers
torch
numpy
gunicorn
google-generativeai
